requirements_txt.txt
# Requirements for Microscopy Image Alignment Tool
# Install with: pip install -r requirements.txt

# Core dependencies
PySide6>=6.5.0
numpy>=1.24.0,<2.0
scipy>=1.10.0
opencv-python>=4.8.0
scikit-image>=0.21.0
tifffile>=2023.7.0
Pillow>=10.0.0

# For creating executable
pyinstaller>=5.13.0